// src/client/pages/Container/Depozitare/Interfete.ts
export interface ContainerInchiriere {
  id_container?: number;
  denumire: string;
  capacitate: number;
  status: number;
  strada: string;
  numar: string;
  latitudine: number;
  longitudine: number;
  localitate: string;
  firma: number;
  denumire_firma: string;
  status_aprobare: number;
  descriere: string;
}